+++ contracts/interface/IRegistry.sol
@@ -2,7 +2,9 @@ pragma solidity ^0.6.0;
 
 
 interface IRegistry {
-    function infos(address) external view returns (bytes32);
+    function handlers(address) external view returns (bytes32);
+    function callers(address) external view returns (bytes32);
 
-    function isValid(address handler) external view returns (bool result);
+    function isValidHandler(address handler) external view returns (bool result);
+    function isValidCaller(address handler) external view returns (bool result);
 }